Use general image dropzone for both
